refactored code for task 3
